requirements.txt
boto3==1.23.10
botocore==1.26.10
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
# dataclasses==0.8
debugpy==1.5.1
filelock==3.4.1
future==1.0.0
huggingface-hub==0.4.0
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
jmespath==0.10.0
joblib==1.1.1
numpy==1.19.5
packaging==21.3
Pillow==8.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-pretrained-bert==0.6.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.27.1
s3transfer==0.5.2
sacremoses==0.0.53
six==1.16.0
tokenizers==0.12.1
# torch @ file:///home/mnt/tangwei3/codes/transCP/torch-1.7.0%2Bcu110-cp36-cp36m-linux_x86_64.whl
# torchvision==0.8.1
tqdm==4.64.1
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.18
zipp==3.6.0
